Extracted `int96RebaseMode`, `datetimeRebaseMode` into `Spark32DataSourceUtils`
